+++ src/main/java/org/example/Modelo/Enemigo/EnemigoTanque.java
@@ -5,7 +5,7 @@ import org.example.Modelo.Posicion;
 import java.util.ArrayList;
 
 public class EnemigoTanque extends EnemigoDeTierra {
-    public EnemigoTanque(int x, int y, ArrayList<Posicion> recorrido){
+    public EnemigoTanque(double x, double y, ArrayList<Posicion> recorrido){
         super(x,y,100,10,recorrido);
     }
